+++ src/app/directives/long-press.directive.ts
@@ -27,7 +27,7 @@ export class LongPressDirective implements OnDestroy {
         this.eventSubscribe = merge(mousedown, mouseup, touchstart, touchEnd)
             .pipe(
                 switchMap((state) => (state ? timer(this.threshold, 1000) : of(null))),
-                filter((value) => value)
+                filter((value) => !!value)
             )
             .subscribe(() => this.mouseLongPress.emit());
     }